결제하기 input기능 구현 완료
최소결제금액 수정
할인금액, 예상충전밤부 출력기능 구현

+++ jPanda/src/main/java/com/kakao/jPanda/kyg/dao/ChargeDao.java
@@ -14,7 +14,7 @@ public interface ChargeDao {
 
 	CouponUseDto selectCouponUse(CouponUseDto couponUseDto);
 
-	Long selectAvailAmountCoupon(CouponUseDto couponUseDto);
+	int selectAvailAmountCoupon(CouponUseDto couponUseDto);
 
 	CouponDto selectCouponByCouponNo(String couponNo);
